fix(storybook): restore native css loading
