+++ .storybook/css-dist.js
@@ -1,5 +1,6 @@
 /**
- * @file Storybook compiled dist CSS loader.
+ * @file Storybook compiled dist CSS side-effect loader.
  */
 
-import 'virtual:emulsify-storybook-css/dist';
+// Import matching CSS for Vite side effects so CSS HMR stays on the native path.
+import.meta.glob('../../../../dist/**/*.css', { eager: true });
