+++ .storybook/main.test.js
@@ -296,131 +296,6 @@ describe('Storybook main config', () => {
     expect(build.emptyOutDir).toBe(false);
   });
 
-  it('loads compiled dist CSS as stylesheet links instead of module imports', async () => {
-    const script = `
-      const { mkdirSync, mkdtempSync, writeFileSync } = await import('node:fs');
-      const { tmpdir } = await import('node:os');
-      const path = await import('node:path');
-      const { pathToFileURL } = await import('node:url');
-
-      const repoRoot = process.cwd();
-      const projectRoot = mkdtempSync(path.join(tmpdir(), 'emulsify-storybook-'));
-      mkdirSync(path.join(projectRoot, 'dist/components/card'), { recursive: true });
-      mkdirSync(path.join(projectRoot, 'dist/global'), { recursive: true });
-      writeFileSync(path.join(projectRoot, 'dist/components/card/card.css'), '.card {}');
-      writeFileSync(path.join(projectRoot, 'dist/global/foundation.css'), '.foundation {}');
-      process.chdir(projectRoot);
-
-      const { default: config } = await import(
-        pathToFileURL(path.join(repoRoot, '.storybook/main.js')).href
-      );
-      const finalConfig = await config.viteFinal({
-        mode: 'development',
-        server: {
-          fs: {
-            allow: [],
-          },
-        },
-        optimizeDeps: {},
-      });
-      const plugin = finalConfig.plugins.find(
-        (item) => item && item.name === 'emulsify-storybook-css-links',
-      );
-
-      console.log(JSON.stringify({
-        dist: plugin.load(
-          plugin.resolveId('virtual:emulsify-storybook-css/dist'),
-        ),
-        sharedDist: plugin.load(
-          plugin.resolveId('virtual:emulsify-storybook-css/shared-dist'),
-        ),
-      }));
-    `;
-    const output = execFileSync(process.execPath, [
-      '--input-type=module',
-      '--eval',
-      script,
-    ]);
-    const modules = JSON.parse(output.toString());
-
-    expect(modules.dist).toContain('"dist/components/card/card.css"');
-    expect(modules.dist).toContain('"dist/global/foundation.css"');
-    expect(modules.dist).toMatch(/document\.createElement\('link'\)/);
-    expect(modules.dist).not.toContain('import.meta.glob');
-    expect(modules.sharedDist).not.toContain('dist/components/card/card.css');
-    expect(modules.sharedDist).toContain('"dist/global/foundation.css"');
-  });
-
-  it('updates dist CSS links without a full Storybook reload', async () => {
-    const script = `
-      const { mkdirSync, mkdtempSync, writeFileSync } = await import('node:fs');
-      const { tmpdir } = await import('node:os');
-      const path = await import('node:path');
-      const { pathToFileURL } = await import('node:url');
-
-      const repoRoot = process.cwd();
-      const projectRoot = mkdtempSync(path.join(tmpdir(), 'emulsify-storybook-'));
-      mkdirSync(path.join(projectRoot, 'dist/global'), { recursive: true });
-      writeFileSync(path.join(projectRoot, 'dist/global/foundation.css'), '.foundation {}');
-      process.chdir(projectRoot);
-
-      const { default: config } = await import(
-        pathToFileURL(path.join(repoRoot, '.storybook/main.js')).href
-      );
-      const finalConfig = await config.viteFinal({
-        mode: 'development',
-        server: {
-          fs: {
-            allow: [],
-          },
-        },
-        optimizeDeps: {},
-      });
-      const plugin = finalConfig.plugins.find(
-        (item) => item && item.name === 'emulsify-storybook-css-links',
-      );
-      const watcherCallbacks = {};
-      const messages = [];
-
-      plugin.configureServer({
-        middlewares: {
-          use() {},
-        },
-        watcher: {
-          add() {},
-          on(eventName, callback) {
-            watcherCallbacks[eventName] = callback;
-          },
-        },
-        ws: {
-          send(message) {
-            messages.push(message);
-          },
-        },
-      });
-
-      watcherCallbacks.change(path.join(projectRoot, 'dist/global/foundation.css'));
-
-      console.log(JSON.stringify(messages));
-    `;
-    const output = execFileSync(process.execPath, [
-      '--input-type=module',
-      '--eval',
-      script,
-    ]);
-    const messages = JSON.parse(output.toString());
-
-    expect(messages).toHaveLength(2);
-    expect(messages.every((message) => message.type === 'custom')).toBe(true);
-    expect(messages.map((message) => message.data.sourceName)).toEqual([
-      'dist',
-      'shared-dist',
-    ]);
-    expect(messages[0].data.stylesheetHrefs).toContain(
-      'dist/global/foundation.css',
-    );
-  });
-
   it('serves generated dist assets that appear after Storybook config loads', async () => {
     const script = `
       const { mkdirSync, mkdtempSync, writeFileSync } = await import('node:fs');
@@ -445,9 +320,10 @@ describe('Storybook main config', () => {
         optimizeDeps: {},
       });
       const plugin = finalConfig.plugins.find(
-        (item) => item && item.name === 'emulsify-storybook-css-links',
+        (item) => item && item.name === 'emulsify-generated-dist-files',
       );
       let middleware;
+      // Capture the middleware directly so the test can exercise late dist files.
       plugin.configureServer({
         middlewares: {
           use(fn) {
@@ -457,13 +333,6 @@ describe('Storybook main config', () => {
         watcher: {
           add() {},
           on() {},
-        },
-        moduleGraph: {
-          getModuleById() {},
-          invalidateModule() {},
-        },
-        ws: {
-          send() {},
         },
       });
 
@@ -526,6 +395,7 @@ describe('Storybook main config', () => {
     ]);
   });
 
+  // The final config must exclude only active Twig virtual modules from prebundling.
   it('dedupes React runtime modules in the final Vite config', async () => {
     const script = `
       const { default: config } = await import('./.storybook/main.js');
@@ -579,8 +449,6 @@ describe('Storybook main config', () => {
       'virtual:emulsify-twig-globs',
       'virtual:emulsify-twig-asset-sources',
       'virtual:emulsify-twig-extension-installers',
-      'virtual:emulsify-storybook-css/dist',
-      'virtual:emulsify-storybook-css/shared-dist',
       '@emulsify/core/storybook/twig/source-function',
       '@emulsify/core/storybook/twig/source',
       '@emulsify/core/storybook/twig/resolver',
