+++ config/jest.config.js
@@ -17,12 +17,11 @@ export default {
     ],
   },
   moduleNameMapper: {
+    // Keep Jest mappings limited to virtual modules still used by Storybook.
     '^virtual:emulsify-twig-globs$':
       '<rootDir>/config/vite/test-utils/virtual-twig-globs.js',
     '^virtual:emulsify-twig-asset-sources$':
       '<rootDir>/config/vite/test-utils/virtual-twig-asset-sources.js',
-    '^virtual:emulsify-storybook-css/(dist|shared-dist)$':
-      '<rootDir>/config/vite/test-utils/empty-module.js',
   },
   coverageDirectory: '.coverage',
   coverageProvider: 'v8',
