--- config/vite/test-utils/empty-module.js
@@ -1,1 +1,0 @@
-export {};
