+++ docs/performance.md
@@ -65,9 +65,7 @@ That fallback blocks the main thread, is deprecated, and is scheduled for remova
 
 ## Storybook CSS Loading
 
-Storybook eagerly loads CSS from the selected render path by default so component styles are available in the iframe. `none` projects load compiled CSS from `dist/**/*.css` as stylesheet links. Drupal projects that mirror component output import component CSS from `components/**/*.css` and load shared compiled CSS from `dist/**/*.css` excluding `dist/components/**/*.css`, because `dist/components` and root `components` represent the same component CSS through different paths.
-
-`dist` CSS is linked instead of module-imported so Storybook's `/dist` static mount can serve the CSS file with the correct MIME type during development.
+Storybook eagerly loads CSS from the selected render path by default so component styles are available in the iframe. `none` projects import compiled CSS from `dist/**/*.css`. Drupal projects that mirror component output import component CSS from `components/**/*.css` and shared compiled CSS from `dist/**/*.css` excluding `dist/components/**/*.css`, because `dist/components` and root `components` represent the same component CSS through different paths.
 
 Projects with very large CSS libraries can opt out and import CSS from their own Storybook preview override:
